Use a standalone resolver for restore.restoreLocation
